Consolidate models.py
