+++ docs/api.md
@@ -1,5 +1,3 @@
 # API reference
 
 ::: iup
-
-::: iup.models
